job search functionality
