+++ frontend/job_portal/src/App.js
@@ -9,7 +9,7 @@ import ForgotPasswordEmail from './Pages/ForgotPasswordEmail'
 import ForgotPasswordOTP from "./Pages/ForgotPasswordOtp";
 import JobPostingPage from "./Pages/Job Posting Page/JobPostingPage"
 import HR_Dashboard from "./Pages/Dashboard/Src/HR_Dashboard";
-import JobSearch from "./Pages/Job Searching Page/jobSearch";
+import JobPortal from './Pages/Job Search/JobPortal'
 
 function App() {
   return (
@@ -32,7 +32,7 @@ function App() {
           <Route path="/forgotpassword" element={<ForgotPassword />} />
           <Route path="/jobpostingpage" element={<JobPostingPage />} />
           <Route path="/hrdashboard" element={<HR_Dashboard />} />
-          <Route path="/jobsearch" element={<JobSearch />} />
+          <Route path="/jobsearch" element={<JobPortal />} />
         </Routes>
       </div>
     </Router>
