+++ frontend/job_portal/src/Pages/Job Search/JobPortal.css
@@ -1,0 +1,240 @@
+/* Full screen container for the job portal */
+.job-portal {
+  width: 100vw;
+  height: 100vh;
+  display: flex;
+  flex-direction: column;
+  background-color: #f3f4f6;
+  overflow: auto;
+}
+
+/* Navbar styling */
+.navbar {
+  width: 100%;
+  padding: 1rem;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  background: linear-gradient(45deg, #7f5cf1, #5a3bed);
+  color: white;
+}
+
+.navbar h1 {
+  margin: 0;
+}
+
+.navbar ul {
+  list-style: none;
+  display: flex;
+  gap: 1rem;
+}
+
+.navbar ul li {
+  cursor: pointer;
+}
+
+.auth-buttons button {
+  margin-left: 0.5rem;
+  padding: 0.5rem 1rem;
+  border: none;
+  cursor: pointer;
+  color: #7f5cf1;
+  background-color: white;
+  border-radius: 5px;
+}
+
+.auth-buttons .sign-up {
+  background-color: #5a3bed;
+  color: white;
+}
+
+/* Search bar styling */
+.search-bar {
+  width: 100%;
+  padding: 1rem 2rem;
+  background: white;
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+  margin: 0;
+  display: flex;
+  justify-content: center;
+}
+
+.search-bar input {
+  width: 100%;
+  max-width: 1200px;
+  padding: 0.75rem;
+  border: 1px solid #ccc;
+  border-radius: 5px;
+  font-size: 1rem;
+}
+
+/* Main content styling */
+.content {
+  display: flex;
+  flex-direction: row;
+  padding: 2rem;
+  gap: 2rem;
+}
+
+/* Job list styling */
+.job-list {
+  flex: 2;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.job-card {
+  width: 100%;
+  max-width: 600px;
+  background: white;
+  padding: 1.5rem;
+  margin: 1rem 0;
+  border-radius: 10px;
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+}
+
+.job-card h2 {
+  margin: 0.5rem 0;
+  font-size: 1.5rem;
+  color: #333;
+}
+
+.job-card h3 {
+  margin: 0;
+  color: #666;
+}
+
+.job-card p {
+  margin: 0.5rem 0;
+}
+
+.skills {
+  margin-top: 1rem;
+}
+
+.skill-badge {
+  display: inline-block;
+  margin-right: 0.5rem;
+  padding: 0.3rem 0.5rem;
+  background: #e0e7ff;
+  color: #5a3bed;
+  border-radius: 5px;
+  font-size: 0.9rem;
+}
+
+.salary {
+  font-weight: bold;
+  color: #5a3bed;
+}
+
+.rating {
+  color: #f9a825;
+}
+
+.verified {
+  font-size: 0.9rem;
+  color: #28a745;
+  font-weight: bold;
+}
+
+/* Selected job details styling */
+.selected-job-card {
+  flex: 1.5;
+  background: white;
+  padding: 1.5rem;
+  border-radius: 10px;
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+  margin: 1rem 0;
+  transition: all 0.3s ease;
+}
+
+.selected-job-card h2 {
+  font-size: 1.5rem;
+  color: #333;
+}
+
+.selected-job-card h3 {
+  color: #666;
+}
+
+.selected-job-card p {
+  margin: 0.5rem 0;
+}
+
+.selected-job-card .skills {
+  margin-top: 0.5rem;
+}
+
+.selected-job-card .skill-badge {
+  display: inline-block;
+  margin-right: 0.5rem;
+  padding: 0.3rem 0.5rem;
+  background: #e0e7ff;
+  color: #5a3bed;
+  border-radius: 5px;
+  font-size: 0.9rem;
+}
+
+.selected-job-card .salary,
+.selected-job-card .rating {
+  font-weight: bold;
+}
+
+.selected-job-card .verified {
+  color: green;
+  font-weight: bold;
+}
+
+/* Profile section styling */
+.profile-card {
+  flex: 1;
+  background: white;
+  padding: 1.5rem;
+  border-radius: 10px;
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+  text-align: center;
+}
+
+.profile-image {
+  width: 100px;
+  height: 100px;
+  border-radius: 50%;
+  margin-bottom: 1rem;
+}
+
+.profile-card h2 {
+  font-size: 1.5rem;
+  color: #333;
+}
+
+.profile-card h3 {
+  color: #666;
+  margin-top: 0.5rem;
+}
+
+.availability {
+  color: #28a745;
+  font-weight: bold;
+  margin-top: 0.5rem;
+}
+
+.profile-stats {
+  display: flex;
+  justify-content: space-around;
+  margin: 1.5rem 0;
+}
+
+.connects,
+.proposals {
+  text-align: center;
+}
+
+.edit-profile {
+  background-color: #7f5cf1;
+  color: white;
+  border: none;
+  padding: 0.5rem 1rem;
+  border-radius: 5px;
+  cursor: pointer;
+}
